backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
httpx>=0.25.0
python-multipart>=0.0.21
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
python-barcode>=0.15.0
pillow>=12.0.0
opencv-python>=4.12.0.88
pyzbar>=0.1.9
numpy>=1.24.0
aiofiles>=23.0.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.23
alembic>=1.13.1
